Rename sourcemaps git file
