+++ src/commands/sourcemaps/common/__tests__/git-sourcemaps-repository-data.test.ts
@@ -1,4 +1,4 @@
-import {getRepositoryData, gitRemote, stripCredentials, TrackedFilesMatcher} from '../git'
+import {getRepositoryData, gitRemote, stripCredentials, TrackedFilesMatcher} from '../git-sourcemaps-repository-data'
 
 describe('git', () => {
   describe('gitRemote', () => {
